# server/swagger.yml
openapi: 3.1.0
info:
  title: PLANT System API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: 本機開發環境
security: []
...
paths:
...
  /api/runs:
    get:
      tags: [Runs]
      summary: 列出執行紀錄
      operationId: listRuns
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
          description: 限定單一專案；未提供時列出可讀取的執行紀錄
      responses:
        "200":
          description: 執行紀錄
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/RunState"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Runs]
      summary: 建立 Agent 執行
      operationId: createRun
      security:
        - ActivationCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCreate"
            example:
              project_id: "173949223214"
              mode: new
              rounds: 2
              rough_idea: 建立訂餐外送系統
              attached_reference_paths: ["173949223214/規定.docx"]
      responses:
        "200":
          description: Run 已排入執行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/runs/{run_id}/events:
    get:
      tags: [Runs]
      summary: 取得 Run 事件或訂閱 SSE
      description: |
        `Accept: application/json` 時回傳 JSON；其他情況以 `text/event-stream` 持續推送事件。
      operationId: getRunEvents
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: since
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 從指定事件 ID 開始讀取
      responses:
        "200":
          description: JSON 事件列表或 SSE 串流
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEventsResponse"
            text/event-stream:
              schema:
                type: string
              example: "event: workspace_event\ndata: {\"id\":1,\"type\":\"run_started\"}\n\n"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/runs/{run_id}/decisions/{decision_id}:
    post:
      tags: [Runs]
      summary: 提交人工決策
      operationId: submitRunDecision
      security:
        - ActivationCookie: []
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/DecisionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionSubmit"
      responses:
        "200":
          description: 更新後的 Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/runs/{run_id}/cancel:
    post:
      tags: [Runs]
      summary: 取消 Run
      operationId: cancelRun
      security:
        - ActivationCookie: []
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: 更新後的 Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunState"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/secrets/model-api-keys:
    get:
      tags: [Secrets]
      summary: 取得模型 API Key 狀態
      description: 不會回傳 API Key 原文。
      operationId: getModelApiKeys
      responses:
        "200":
          description: Provider 狀態
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/ModelApiKeyStatus"
    put:
      tags: [Secrets]
      summary: 儲存模型 API Key
      operationId: updateModelApiKey
      security:
        - ActivationCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelApiKeyUpdate"
            example:
              provider: openai
              api_key: sk-example
      responses:
        "200":
          description: API Key 已儲存並重設為未測試
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiKeyStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/secrets/model-api-keys/test:
    post:
      tags: [Secrets]
      summary: 測試模型 API Key
      operationId: testModelApiKey
      security:
        - ActivationCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelApiKeyTest"
      responses:
        "200":
          description: 測試結果；無效 Key 仍以 200 回傳並將 valid 設為 false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiKeyStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/secrets/model-api-keys/{provider}:
    delete:
      tags: [Secrets]
      summary: 刪除模型 API Key
      operationId: deleteModelApiKey
      security:
        - ActivationCookie: []
      parameters:
        - $ref: "#/components/parameters/Provider"
      responses:
        "200":
          description: API Key 刪除結果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ModelApiKeyStatus"
                  - type: object
                    properties:
                      deleted: { type: boolean }
                      removed: { type: boolean }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/secrets/activation-code:
    get:
      tags: [Secrets]
      summary: 取得啟動狀態
      operationId: getActivationStatus
      responses:
        "200":
          description: 啟動狀態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationStatus"
    post:
      tags: [Secrets]
      summary: 使用啟動碼啟動寫入權限
      operationId: activate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCodePayload"
      responses:
        "200":
          description: 啟動成功並設定 HttpOnly Cookie
          headers:
            Set-Cookie:
              schema: { type: string }
              description: plant_activation Cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Secrets]
      summary: 移除啟動 Cookie
      operationId: deactivate
      responses:
        "200":
          description: 已移除啟動 Cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationStatus"
...
components:
  securitySchemes:
    ActivationCookie:
      type: apiKey
      in: cookie
      name: plant_activation
      description: 透過 POST /api/secrets/activation-code 取得的 HttpOnly Cookie
  parameters:
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string, pattern: "^run_[A-Za-z0-9]+$" }
    DecisionId:
      name: decision_id
      in: path
      required: true
      schema: { type: string }
...
    Provider:
      name: provider
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Provider"
...
  responses:
    BadRequest:
      description: 請求資料或操作條件不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: 缺少啟動權限或專案目前不可公開讀取
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { detail: 需要啟動碼才能執行此操作 }
    NotFound:
      description: 找不到指定資源
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: 資源狀態衝突，例如專案已有執行中的 Run
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooManyRequests:
      description: 啟動碼嘗試次數過多，請依 Retry-After 稍後重試
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
          description: 可再次嘗試前需等待的秒數
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: FastAPI/Pydantic 請求格式驗證失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
...
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          description: 錯誤訊息或詳細資料
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            additionalProperties: true
...
    RunStatus:
      type: string
      enum:
        - queued
        - running
        - waiting_for_human
        - cancelling
        - completed
        - failed
        - cancelled
        - interrupted
...
    RunCreate:
      type: object
      required: [project_id]
      properties:
        project_id: { type: string }
        mode: { type: string, enum: [new, continue], default: continue }
        rounds: { type: [integer, "null"], minimum: 0, maximum: 20 }
        max_issues: { type: [integer, "null"], minimum: 1, maximum: 50 }
        rough_idea: { type: [string, "null"] }
        attached_reference_paths:
          type: [array, "null"]
          items: { type: string }
        enable_agents:
          type: [object, "null"]
          additionalProperties: { type: boolean }
        stage_overrides:
          type: [object, "null"]
          additionalProperties: { type: boolean }
    RunState:
      type: object
      required: [run_id, project_id, mode, status]
      properties:
        run_id: { type: string }
        project_id: { type: string }
        mode: { type: string, enum: [new, continue] }
        status: { $ref: "#/components/schemas/RunStatus" }
        current_stage: { type: string }
        current_agent: { type: string }
        round: { type: [integer, "null"] }
        rough_idea: { type: string }
        attached_reference_paths:
          type: array
          items: { type: string }
        requires_rounds_input: { type: boolean }
        pending_decision:
          oneOf:
            - $ref: "#/components/schemas/PendingDecision"
            - type: "null"
        skip_all_human_interventions: { type: boolean }
        cancel_requested: { type: boolean }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        error: { type: [string, "null"] }
        event_count: { type: integer, minimum: 0 }
        run_checkpoint:
          oneOf:
            - $ref: "#/components/schemas/RunCheckpoint"
            - type: "null"
      additionalProperties: true
    RunCheckpoint:
      type: object
      required: [status, stage_id, run_id]
      properties:
        status: { type: string }
        stage_id: { type: string }
        step_id: { type: string }
        run_id: { type: string }
        error: { type: string }
        dirty_outputs:
          type: array
          items: { type: string }
        last_round: { type: integer }
        round: { type: integer }
        issue_id: { type: string }
        agent: { type: string }
        action: { type: string }
        resume_policy: { type: string }
        created_at: { type: string, format: date-time }
    PendingDecision:
      type: object
      required: [id, kind, title, description]
      properties:
        id: { type: string }
        kind:
          type: string
          enum:
            - stakeholder_selection
            - human_decision
            - stakeholder_statement_review
            - requirements_review
            - domain_research_review
            - scope_review
            - meeting_issue_proposal_review
        title: { type: string }
        description: { type: string }
        proposed:
          type: array
          items:
            type: object
            additionalProperties: true
        max_select: { type: integer }
        issue:
          type: object
          additionalProperties: true
        options:
          type: object
          additionalProperties: true
    DecisionSubmit:
      type: object
      required: [payload]
      properties:
        payload:
          type: object
          additionalProperties: true
      example:
        payload:
          action: approve
    RunEvent:
      type: object
      required: [id, type]
      properties:
        id: { type: integer, minimum: 0 }
        type: { type: string }
        message: { type: string }
        level: { type: string }
        timestamp: { type: string, format: date-time }
        stage_id: { type: string }
        step_id: { type: string }
        agent: { type: string }
        action: { type: string }
        title: { type: string }
        status: { type: string }
        output_path: { type: string }
        delta_type: { type: string }
        content: true
        summary:
          type: object
          additionalProperties: true
        decision_id: { type: string }
        decision: { $ref: "#/components/schemas/PendingDecision" }
        payload:
          type: object
          additionalProperties: true
        attached_reference_paths:
          type: array
          items: { type: string }
        error: { type: string }
      additionalProperties: true
    RunEventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/RunEvent" }
...
    Provider:
      type: string
      enum: [openai, claude, gemini]
    ApiKeyValidationStatus:
      type: string
      enum: [untested, valid, invalid]
    ModelApiKeyStatus:
      type: object
      required: [provider, env_key, configured, status, valid]
      properties:
        provider: { $ref: "#/components/schemas/Provider" }
        env_key: { type: string }
        configured: { type: boolean }
        status: { $ref: "#/components/schemas/ApiKeyValidationStatus" }
        valid: { type: boolean }
        error: { type: [string, "null"] }
        tested_at: { type: [string, "null"], format: date-time }
      additionalProperties: true
    ModelApiKeyUpdate:
      type: object
      required: [provider, api_key]
      properties:
        provider: { $ref: "#/components/schemas/Provider" }
        api_key:
          type: string
          minLength: 1
          writeOnly: true
    ModelApiKeyTest:
      type: object
      required: [provider]
      properties:
        provider: { $ref: "#/components/schemas/Provider" }
    ActivationCodePayload:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          writeOnly: true
    ActivationStatus:
      type: object
      required: [activated]
      properties:
        activated: { type: boolean }
...